requirements.txt
apache-airflow-providers-amazon==8.25.0
Faker==26.0.0
pandas==2.2.2
